requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2

